+++ lib/projects/utils.ts
@@ -1,31 +1,3 @@
-// import {PROJECTS} from '@/config/exercises'
-// import { Exercise } from '@/types/utils'
-
-// export type ExerciseCollection = {
-//   [key: string]: Exercise
-// }
-
-// export function getExerciseById(id: string): Exercise | undefined {
-//   return Object.values(PROJECTS).find((exercise) => exercise.id === id)
-// }
-
-// export function getPublishedExercises(): Exercise[] {
-//   return Object.values(PROJECTS).filter((exercise) => exercise.published)
-// }
-
-// export function getExercisesByDifficulty(
-//   difficulty: Exercise['difficulty']
-// ): Exercise[] {
-//   return Object.values(PROJECTS).filter(
-//     (exercise) => exercise.published && exercise.difficulty === difficulty
-//   )
-// }
-
-// export function getExercisesByTag(tag: string): Exercise[] {
-//   return Object.values(PROJECTS).filter(
-//     (exercise) => exercise.published && exercise.tags.includes(tag)
-//   )
-// }
 import { Project, ProjectCollection } from '@/types/utils'
 import { PROJECTS } from '@/config/projects'
